Minor bug correction in mni\result\f06\read_disp_complex.m
I think line 52 should be "if count == 6" not "if count == 7" based on how my f06s print (for a continuous turbulence 146 analysis in NASTRAN 2023). NASTRAN appears to print the imaginary displacement components as six floats with nothing else on the same line (i.e. no random zero in the first column). Thus scanning for six floats reads the line correctly.